chore: move config files from config to security and swagger
